Moved elevator pitch as its own tab
