+++ src/app/workshop/[id]/page.tsx
@@ -5,10 +5,12 @@ import { useParams, useRouter } from 'next/navigation';
 import { Button } from "@/components/ui/button";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { useWorkshop } from '@/lib/workshop-context';
+import { useEvaluations } from '@/hooks/useEvaluations';
 import { CardSelection } from '@/components/workshop/card-selection';
 import { IdeaRefinement } from '@/components/workshop/idea-refinement';
 import { Storyboard } from '@/components/workshop/storyboard';
 import { Evaluation } from '@/components/workshop/evaluation';
+import { ElevatorPitch } from '@/components/workshop/elevator-pitch';
 import { WorkshopPhase, Idea, CardType } from '@/types';
 import { toast } from 'sonner';
 
@@ -97,6 +99,7 @@ const IdeasList = ({
 
 export default function WorkshopPage() {
   const workshop = useWorkshop();
+  const { selectedCriteria } = useEvaluations();
   const { 
     currentWorkshop, 
     setCurrentWorkshop, 
@@ -114,7 +117,7 @@ export default function WorkshopPage() {
   
   // Add isEditingIdea state to track when we're editing vs creating
   const [isEditingIdea, setIsEditingIdea] = useState(false);
-  
+
   // Load the workshop
   useEffect(() => {
     // Extract the workshop ID from the URL
@@ -266,7 +269,7 @@ export default function WorkshopPage() {
   const handlePhaseChange = (phase: string) => {
     setCurrentPhase(phase as WorkshopPhase);
   };
-  
+
   return (
     <div className="container mx-auto py-8 px-4 max-w-7xl">
       <div className="mb-8 space-y-4">
@@ -300,7 +303,7 @@ export default function WorkshopPage() {
         {/* Main content area */}
         <div className="lg:col-span-3">
           <Tabs value={currentPhase} onValueChange={handlePhaseChange}>
-            <TabsList className="grid grid-cols-4 w-full mb-8">
+            <TabsList className="grid grid-cols-5 w-full mb-8">
               <TabsTrigger value="ideation">
                 <div className="flex items-center gap-2">
                   <span className="bg-primary text-white w-6 h-6 rounded-full flex items-center justify-center text-xs">1</span>
@@ -325,6 +328,12 @@ export default function WorkshopPage() {
                   Evaluation
                 </div>
               </TabsTrigger>
+              <TabsTrigger value="elevator" disabled={!currentIdea}>
+                <div className="flex items-center gap-2">
+                  <span className="bg-primary text-white w-6 h-6 rounded-full flex items-center justify-center text-xs">5</span>
+                  Elevator Pitch
+                </div>
+              </TabsTrigger>
             </TabsList>
             
             <TabsContent value="ideation">
@@ -345,6 +354,10 @@ export default function WorkshopPage() {
             
             <TabsContent value="evaluation">
               <Evaluation />
+            </TabsContent>
+
+            <TabsContent value="elevator">
+              <ElevatorPitch />
             </TabsContent>
           </Tabs>
         </div>
